third_party/blink/web_tests/platform/generic/fast/frames/open-then-unload-expected.txt

Test that document.open() doesn't allow frames to be attached during frame detach.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS 1 is window[0].length
PASS window[0][1] is undefined.
PASS successfullyParsed is true

TEST COMPLETE
